Codechange: Signed value silently converted to unsigned and back.

+++ src/disaster_vehicle.cpp
@@ -463,7 +463,7 @@ static bool DisasterTick_Aircraft(DisasterVehicle *v, uint16_t image_override, b
 			if (_settings_client.sound.disaster) SndPlayTileFx(SND_12_EXPLOSION, i->location.tile);
 		}
 	} else if (v->state == 0) {
-		int x = v->x_pos + ((leave_at_top ? -15 : 15) * TILE_SIZE);
+		int x = v->x_pos + ((leave_at_top ? -15 : 15) * (int)TILE_SIZE);
 		int y = v->y_pos;
 
 		if ((uint)x > Map::MaxX() * TILE_SIZE - 1) return true;
